Develop

+++ backend/controllers/cards.js
@@ -1,58 +1,68 @@
 const { default: mongoose } = require('mongoose');
 const Cards = require('../models/card');
-
-const ERROR_CODE = 400;
-const NOT_FOUND = 404;
-const SERVEL_ERROR = 500;
+const { NOT_FOUND, SERVER_ERROR, FORBIDDEN } = require('../utils/constants');
 
 module.exports.getCards = async (req, res) => {
   try {
-    const cards = await Cards.find({}).orFail();
+    const cards = await Cards.find({}).sort({ createdAt: -1 }).orFail();
 
     return res.json({ cards });
   } catch (err) {
     console.error(err);
     return res
-      .status(SERVEL_ERROR)
+      .status(SERVER_ERROR)
       .json({ message: 'Error interno del servidor' });
   }
 };
 
 module.exports.createCard = async (req, res) => {
   const { name, link } = req.body;
-  const ownerId = req.headers['user'];
+  const ownerId = req.user._id;
 
   if (!name || !link) {
-    return res
-      .status(ERROR_CODE)
-      .json({ message: 'Información no encontrado' });
+    return res.status(NOT_FOUND).json({ message: 'Información no encontrado' });
   }
 
   try {
-    await Cards.create({ name, link, owner: ownerId });
+    const newCard = await Cards.create({ name, link, owner: ownerId });
 
-    return res.status(201).json({ name, link, owner: ownerId });
+    return res
+      .status(201)
+      .json({ _id: newCard._id.toString(), name, link, owner: ownerId });
   } catch (err) {
     console.error(err);
 
     if (err.name === 'ValidationError') {
       return res
-        .status(ERROR_CODE)
+        .status(NOT_FOUND)
         .json({ message: 'Datos de tarjeta invalidos' });
     }
 
     return res
-      .status(SERVEL_ERROR)
+      .status(SERVER_ERROR)
       .json({ message: 'Error al crear una nueva tarjeta' });
   }
 };
 
 module.exports.deleteCard = async (req, res) => {
   const cardId = req.params.cardId;
+  const userId = req.user._id;
 
   try {
     if (!mongoose.Types.ObjectId.isValid(cardId)) {
-      return res.status(ERROR_CODE).json({ message: 'ID de tarjeta invalido' });
+      return res.status(NOT_FOUND).json({ message: 'ID de tarjeta invalido' });
+    }
+
+    const card = await Cards.findById(cardId);
+
+    if (!card) {
+      return res.status(NOT_FOUND).json({ message: 'Tarjeta no encontrada' });
+    }
+
+    if (card.owner.toString() !== userId) {
+      return res
+        .status(FORBIDDEN)
+        .json({ message: 'No tienes permiso para borrar esta tarjeta' });
     }
 
     await Cards.findByIdAndDelete(cardId).orFail();
@@ -66,7 +76,7 @@ module.exports.deleteCard = async (req, res) => {
     }
 
     return res
-      .status(SERVEL_ERROR)
+      .status(SERVER_ERROR)
       .json({ message: 'Error interno del servidor' });
   }
 };
@@ -77,7 +87,7 @@ module.exports.likeCard = async (req, res) => {
 
     if (!mongoose.Types.ObjectId.isValid(userId)) {
       return res
-        .status(ERROR_CODE)
+        .status(FORBIDDEN)
         .json({ message: ' ID de usuario no valido' });
     }
 
@@ -87,12 +97,14 @@ module.exports.likeCard = async (req, res) => {
       { new: true },
     ).orFail();
 
-    return res.status(200).json({ message: 'Le diste like a la card' });
+    Cards.findById(req.params.cardId).then((card) => {
+      res.status(200).json(card);
+    });
   } catch (err) {
     console.error(err);
 
     return res
-      .status(SERVEL_ERROR)
+      .status(SERVER_ERROR)
       .json({ message: 'Error interno del servidor' });
   }
 };
@@ -102,9 +114,7 @@ module.exports.disLikeCard = async (req, res) => {
     const userId = req.user._id;
 
     if (!mongoose.Types.ObjectId.isValid(userId)) {
-      return res
-        .status(ERROR_CODE)
-        .json({ message: 'ID de usuario no valido' });
+      return res.status(NOT_FOUND).json({ message: 'ID de usuario no valido' });
     }
 
     await Cards.findByIdAndUpdate(
@@ -112,13 +122,14 @@ module.exports.disLikeCard = async (req, res) => {
       { $pull: { likes: userId } },
       { new: true },
     ).orFail();
-
-    return res.status(200).json({ message: 'Like removido con exito' });
+    Cards.findById(req.params.cardId).then((card) => {
+      res.status(200).json(card);
+    });
   } catch (err) {
     console.error(err);
 
     return res
-      .status(SERVEL_ERROR)
+      .status(SERVER_ERROR)
       .json({ message: 'Error interno del servidor' });
   }
 };
